pyproject.toml
[tool.poetry]
name = "auto-release"
version = "0.1.0"
description = ""
authors = ["himkt <himkt@cookpad.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.47"

[tool.poetry.dev-dependencies]
python-language-server = "^0.31.8"
ipython = "^7.13.0"
pytest = "^5.4.1"
flake8 = "^3.7.9"
rope = "^0.16.0"
mypy = "^0.770"
black = "^19.10b0"
isort = "^5.5.3"

[tool.black]
line-length = 99

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
